1. 增加了用户获取信息、签到、上传头像等API 2. 修复了日志错误的问题 3. 增加了10分钟自动清理验证码

+++ common/src/main/java/com/virus/pt/common/util/JwtUtils.java
@@ -40,7 +40,7 @@ public class JwtUtils {
      * @return
      * @throws TipException
      */
-    public static Integer getUserIdFromRequest(HttpServletRequest request) throws TipException {
+    public static Long getUserIdFromRequest(HttpServletRequest request) throws TipException {
         if (request == null) {
             throw new TipException(ResultEnum.ARGS_ERROR);
         }
@@ -49,12 +49,12 @@ public class JwtUtils {
         return getUserId(token);
     }
 
-    public static Integer getUserId(String token) throws TipException {
+    public static Long getUserId(String token) throws TipException {
         if (StringUtils.isBlank(token)) {
             throw new TipException(ResultEnum.TOKEN_EMPTY_ERROR);
         }
         try {
-            return JWT.decode(token).getClaim(ApiConst.JWT_CLAIM_USER_ID).asInt();
+            return JWT.decode(token).getClaim(ApiConst.JWT_CLAIM_USER_ID).asLong();
         } catch (JWTDecodeException e) {
             throw new TipException(ResultEnum.TOKEN_DECODE_ERROR);
         }
